Gettting Date and Time stuff
this code helps you to get Date and Time.
As we all can make an idea that time intervals have to be represented in floating point number. The floating-point numbers in units of seconds are signified by Tick in python. Particular instants of time are represented in seconds since 12:00 am, 1st of January, of the year 1990. A popular module name 'time' is available which provides functions that let's programmer work with time and made conversion between time-representation possible